% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cleanUp}
\alias{cleanUp}
\alias{doClust}
\alias{lply}
\alias{getResult}
\title{Clean up for temp files}
\usage{
cleanUp(user = NULL, conf = NULL, host = "login.gbar.dtu.dk",
  keyPath = NULL)
}
\arguments{
\item{user}{server to read from}

\item{host}{print file to read}

\item{keyPath}{}
}
\value{
TRUE/1 if found and deleted
}
\description{
Clean up for temp files
}
