% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{doClust}
\alias{doClust}
\title{do.call with cluster as backend via ssh}
\usage{
doClust(what, arg = list(), user = "sowe", host = "login.gbar.dtu.dk",
  packages = c(), Rscript = TRUE)
}
\arguments{
\item{what}{function or name of function to execute}

\item{arg}{arguments for function}

\item{user}{username}

\item{host}{server (will connect to user@server)}

\item{packages}{char vector of package names (can be empty)}

\item{Rscript}{execute by Rscript or R CMD BATCH (former  only supported on gbar, ladder no verbose)}
}
\value{
value return by evaluated function
}
\description{
do.call with cluster as backend via ssh
}

