% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVfunctions.R
\name{doOneGrid1}
\alias{doOneGrid1}
\title{Perform one job in a grid}
\usage{
doOneGrid1(one, parent.function.name = "model.function", Dim = 1,
  verbose = F, ...)
}
\arguments{
\item{one}{one job, as defined by makeGrid}

\item{parent.function.name}{doOneGrid apply a parent function of this name}

\item{Dim}{is training set X a list dim=1, or data.frame dim=2}

\item{...}{other args passed to model.function}
}
\value{
result from model.function
}
\description{
Perform one job in a grid
}
