% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{exchangeRSAkeys}
\alias{exchangeRSAkeys}
\title{Clean up for temp files}
\usage{
exchangeRSAkeys(user, host, server.ssh.pub.file = "~/.ssh/authorized_keys",
  keyName = "fastRditijuu")
}
\arguments{
\item{user}{server to read from}

\item{host}{print file to read}

\item{host}{print file to read}
}
\value{
TRUE/1 if found and deleted
}
\description{
Clean up for temp files
}

