% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getResult}
\alias{getResult}
\title{Read Rout, useful if Rscript=FALSE to retrieve print}
\usage{
getResult(ticket, conf = NULL, user = NULL, host = "login.gbar.dtu.dk",
  transferHost = NULL, verbose = F, keyPath = NULL)
}
\arguments{
\item{ticket}{the out object from lply or doClust when call with async=TRUE}

\item{user}{username on server}

\item{host}{server host address}

\item{verbose}{if TRUE, prints from R master, can be used to monitor progress and debugging}

\item{transferString}{server to handle file transfer, host or transfer.gbar.dtu.dk if NULL}
}
\value{
Either the result of a finished job, or if not complete yet, then a ticket like object
}
\description{
Read Rout, useful if Rscript=FALSE to retrieve print
}
