% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lply}
\alias{lply}
\title{Powerful lapply function with a qsub(Torque/PBS)-cluster as backend}
\usage{
lply(X, FUN, user = "sowe", host = "login.gbar.dtu.dk", Rscript = T,
  packages = c(), max.nodes = 24, local = FALSE, ...)
}
\arguments{
\item{X}{object to map}

\item{FUN}{mapper function}

\item{user}{username for server}

\item{host}{server adress (will be combined as user@host)}

\item{Rscript}{if true verbose, however only supported by gbar, nor compute.cluster}

\item{packages}{required packages on server}

\item{max.nodes}{maximum nodes reuired, do not set higher than 80}

\item{local}{should the lply run on local computer (only for debugging/testing)}

\item{...}{}
}
\value{
list of results
}
\description{
Powerful lapply function with a qsub(Torque/PBS)-cluster as backend
}

