% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVfunctions.R
\name{makeGrid}
\alias{makeGrid}
\alias{gridSearch}
\title{make a job list for a repeated kfold CV grid search}
\usage{
makeGrid(N, reps = 10, folds = 10, trainArgs = list())
}
\arguments{
\item{N}{number of observations in training set, or a target vector of N elements}

\item{reps}{number of repetitions}

\item{folds}{number of folds}

\item{trainArgs}{named list of model arguments to span a grid over}
}
\value{
a mappable job list, run with lapply or lply
}
\description{
make a job list for a repeated kfold CV grid search
}
\examples{
grid = makeGrid(N=500,trainArgs = list(mtry=1:5,ntree=1:10))
}
