% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readPrint}
\alias{readPrint}
\title{Read Rout, useful if Rscript=FALSE to retrieve print}
\usage{
readPrint(hostString = "sowe@login.gbar.dtu.dk",
  printFile = "runOnServer.Rout")
}
\arguments{
\item{hostString}{server to read from}

\item{printFile}{print file to read}
}
\value{
lines of print
}
\description{
Read Rout, useful if Rscript=FALSE to retrieve print
}

