% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVfunctions.R
\name{repfold}
\alias{repfold}
\alias{gridSearch}
\title{Make indexes for a repeated k-fold cross-validation.}
\usage{
repfold(N, reps = 5, folds = 20)
}
\arguments{
\item{N}{number of observations in training set, or a target vector of N elements}

\item{reps}{number of repetitions}

\item{folds}{number of folds}
}
\value{
a named list of test index for each fold and repetition
}
\description{
Make indexes for a repeated k-fold cross-validation.
}
\examples{
repfold(N=500,reps=10,folds=10)
}
