% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cleanUp}
\alias{cleanUp}
\title{Clean up for temp files}
\usage{
cleanUp(user, host = "login.gbar.dtu.dk")
}
\arguments{
\item{user}{server to read from}

\item{host}{print file to read}
}
\value{
TRUE/1 if found and deleted
}
\description{
Clean up for temp files
}

