% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{doBatchJob}
\alias{doBatchJob}
\title{Internal backend function to perform a qsub batchjob on server}
\usage{
doBatchJob(X, FUN, packages = c(), max.nodes = 24, ...)
}
\value{
list with results of qsub batchjob
}
\description{
Internal backend function to perform a qsub batchjob on server
}

