% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getResult}
\alias{getResult}
\title{Read Rout, useful if Rscript=FALSE to retrieve print}
\usage{
getResult(ticket, user, host = "login.gbar.dtu.dk", verbose = F)
}
\arguments{
\item{vebose}{}
}
\description{
Read Rout, useful if Rscript=FALSE to retrieve print
}

