% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lply}
\alias{lply}
\title{Powerful lapply function with a qsub(Torque/PBS)-cluster as backend}
\usage{
lply(X, FUN, conf = NULL, user = NULL, host = "login.gbar.dtu.dk",
  transferHost = NULL, keyPath = NULL, Rscript = T, packages = c(),
  max.nodes = 8, local = FALSE, globalVar = list(), async = T,
  nCores = 4, qsub.walltime = "00:09:59", qsub.proc = 4, qsub.nodes = 1,
  ...)
}
\arguments{
\item{X}{object to map}

\item{FUN}{mapper function}

\item{user}{username for server}

\item{host}{server adress (will be combined as user@host)}

\item{keyPath}{path to local ssh key (mainly for windows)}

\item{Rscript}{if true verbose, however only supported by gbar, nor compute.cluster}

\item{packages}{required packages on server}

\item{max.nodes}{maximum nodes reuired, do not set higher than 80 (79 if async=T)}

\item{local}{should the lply run on local computer (only for debugging/testing)}

\item{globalVar}{list of global variables}

\item{async}{true/false, if to run lply loop asynchronously}

\item{nCores}{number of cores(>1 will trigger parallel::mclapply)}

\item{qsub.walltime}{change wall time qsub, will apply to both master and slaves}

\item{qsub.proc}{how many processes to ask for per job, 1 unless using doParallel etc.}

\item{qsub.nodes}{how many nodes to ask for per job, leave unchanged if in doubt}

\item{...}{}
}
\value{
list of results
}
\description{
Powerful lapply function with a qsub(Torque/PBS)-cluster as backend
}
